+++ src/Backends/Rendering/Window/Software.h
@@ -2,9 +2,9 @@
 
 #include <stddef.h>
 
-bool WindowBackend_Software_CreateWindow(const char *window_title, int screen_width, int screen_height, bool fullscreen);
+bool WindowBackend_Software_CreateWindow(const char *window_title, size_t screen_width, size_t screen_height, bool fullscreen);
 void WindowBackend_Software_DestroyWindow(void);
 unsigned char* WindowBackend_Software_LockFramebuffer(size_t *pitch);
 void WindowBackend_Software_UnlockFramebuffer(void);
 void WindowBackend_Software_Display(void);
-void WindowBackend_Software_HandleWindowResize(unsigned int width, unsigned int height);
+void WindowBackend_Software_HandleWindowResize(size_t width, size_t height);
