+++ src/Backends/Rendering/Window/Software/WiiU.cpp
@@ -48,7 +48,7 @@ static GX2Texture screen_texture;
 static Viewport tv_viewport;
 static Viewport drc_viewport;
 
-static void CalculateViewport(unsigned int actual_screen_width, unsigned int actual_screen_height, Viewport *viewport)
+static void CalculateViewport(size_t actual_screen_width, size_t actual_screen_height, Viewport *viewport)
 {
 	if (actual_screen_width * fake_framebuffer_height > fake_framebuffer_width * actual_screen_height) // Fancy way to do `if (actual_screen_width / actual_screen_height > fake_framebuffer_width / fake_framebuffer_height)` without floats
 	{
@@ -68,7 +68,7 @@ static void CalculateViewport(unsigned int actual_screen_width, unsigned int act
 	}
 }
 
-bool WindowBackend_Software_CreateWindow(const char *window_title, int screen_width, int screen_height, bool fullscreen)
+bool WindowBackend_Software_CreateWindow(const char *window_title, size_t screen_width, size_t screen_height, bool fullscreen)
 {
 	(void)window_title;
 	(void)fullscreen;
@@ -265,7 +265,7 @@ ATTRIBUTE_HOT void WindowBackend_Software_Display(void)
 	WHBGfxFinishRender();
 }
 
-void WindowBackend_Software_HandleWindowResize(unsigned int width, unsigned int height)
+void WindowBackend_Software_HandleWindowResize(size_t width, size_t height)
 {
 	(void)width;
 	(void)height;
